refactor(app): begin atomic design structure by restructuring App as an environment
Begin the process of converting to atomic design by moving the App component into an Environment
directory
